requirements.txt
transformers==4.40.2
torch==2.2.2
torchvision==0.17.2
gradio==4.26.0
pillow==10.3.0